requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
astroid==4.0.3
asyncer==0.0.12
attrs==25.4.0
audioop-lts==0.2.2
autoevals==0.0.130
beautifulsoup4==4.14.3
bidict==0.23.1
blis==1.3.3
boltons==21.0.0
bracex==2.6
braintrust==0.4.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
chainlit==2.9.4
charset-normalizer==3.4.4
chevron==0.14.0
claude-agent-sdk==0.1.19
click==8.1.8
click-option-group==0.5.9
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
ctranslate2==4.6.3
cuid==0.4
cymem==2.0.13
dataclasses-json==0.6.7
deep-translator==1.11.4
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
emoji==2.15.0
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.109.0
fastuuid==0.14.0
filelock==3.20.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
glom==22.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.133.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.8.0
google-cloud-translate==3.24.0
google-crc32c==1.8.0
google-genai==1.59.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==1.2.7
langchain-google-genai==4.1.3
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.6.2
Lazify==0.4.0
librt==0.7.8
limits==5.6.0
litellm==1.80.16
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mccabe==0.7.0
mcp==1.23.3
mcp-code-checker @ git+https://github.com/MarcusJellinghaus/mcp-code-checker.git@3ef6a374b09e6680c120f1b5f8422d713f74dc40
mcp-config @ git+https://github.com/MarcusJellinghaus/mcp-config.git@5d50065cdfc1dc4648bc120924246a519810aff3
mdurl==0.1.2
motor==3.3.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
neo4j==6.0.3
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
openai==2.15.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-agno==0.50.1
opentelemetry-instrumentation-alephalpha==0.50.1
opentelemetry-instrumentation-anthropic==0.50.1
opentelemetry-instrumentation-bedrock==0.50.1
opentelemetry-instrumentation-chromadb==0.50.1
opentelemetry-instrumentation-cohere==0.50.1
opentelemetry-instrumentation-crewai==0.50.1
opentelemetry-instrumentation-google-generativeai==0.50.1
opentelemetry-instrumentation-groq==0.50.1
opentelemetry-instrumentation-haystack==0.50.1
opentelemetry-instrumentation-lancedb==0.50.1
opentelemetry-instrumentation-langchain==0.50.1
opentelemetry-instrumentation-llamaindex==0.50.1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.50.1
opentelemetry-instrumentation-mcp==0.50.1
opentelemetry-instrumentation-milvus==0.50.1
opentelemetry-instrumentation-mistralai==0.50.1
opentelemetry-instrumentation-ollama==0.50.1
opentelemetry-instrumentation-openai==0.50.1
opentelemetry-instrumentation-openai-agents==0.50.1
opentelemetry-instrumentation-pinecone==0.50.1
opentelemetry-instrumentation-qdrant==0.50.1
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.50.1
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.50.1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.50.1
opentelemetry-instrumentation-transformers==0.50.1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.50.1
opentelemetry-instrumentation-watsonx==0.50.1
opentelemetry-instrumentation-weaviate==0.50.1
opentelemetry-instrumentation-writer==0.50.1
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
peewee==3.19.0
platformdirs==4.5.1
pluggy==1.6.0
polyleven==0.9.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
pymongo==4.6.1
pyparsing==3.3.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.0
python-json-logger==4.0.0
python-multipart==0.0.21
python-slugify==8.0.4
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
sacremoses==0.1.1
scikit-learn==1.8.0
scipy==1.17.0
semgrep==1.148.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.1.2
sseclient-py==1.9.0
stanza==1.10.1
starlette==0.52.1
structlog==25.5.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.0.2
tomlkit==0.14.0
torch==2.9.1
tqdm==4.67.1
traceloop-sdk==0.50.1
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.13.0
uvicorn==0.40.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
wcmatch==8.5.2
weasel==0.4.3
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
